fix - http://jira.magnolia.info/browse/MAGNOLIA-918
git-svn-id: file:///home/migrator/subversion-migration/svn-repo-copy/magnolia/trunk@3400 a15fbcf9-bee4-0310-bafe-d02d26ad3b5f

+++ magnolia-exchange-simple/src/main/java/info/magnolia/cms/exchange/simple/SimpleExchangeServlet.java
@@ -194,7 +194,11 @@ public class SimpleExchangeServlet extends HttpServlet {
         Iterator nodeDataIterator = destination.getNodeDataCollection().iterator();
         while (nodeDataIterator.hasNext()) {
             NodeData nodeData = (NodeData) nodeDataIterator.next();
-            nodeData.delete();
+            // Ignore binary types, since these are sub nodes and already taken care of while
+            // importing sub resources
+            if (nodeData.getType() != PropertyType.BINARY) {
+                nodeData.delete();
+            }
         }
 
         // copy all properties
